Logged in User Acess Updated

+++ src/main/java/org/test/crash_course_springboot/services/UserService.java
@@ -1,6 +1,7 @@
 package org.test.crash_course_springboot.services;
 
 import lombok.extern.slf4j.Slf4j;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -9,17 +10,21 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.test.crash_course_springboot.dto.UserDto;
 import org.test.crash_course_springboot.entities.UserEntity;
+import org.test.crash_course_springboot.exceptions.DuplicateResourceException;
 import org.test.crash_course_springboot.exceptions.ResourceNotFoundException;
 import org.test.crash_course_springboot.exceptions.UnauthorizedActionException;
 import org.test.crash_course_springboot.repo.UserRepo;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
 public class UserService {
+
     @Autowired
     private UserRepo userRepo;
+
     @Autowired
     private PasswordEncoder passwordEncoder;
 
@@ -30,9 +35,10 @@ public class UserService {
         user.setName(dto.getFirstName() + " " + dto.getLastName());
         user.setEmail(dto.getEmail());
         user.setRole(dto.getRole());
-        user.setPassword(dto.getPassword());
+        user.setPassword(dto.getPassword()); // encoded later
         return user;
     }
+
 
     private UserDto entityToDto(UserEntity user) {
         UserDto dto = new UserDto();
@@ -48,72 +54,78 @@ public class UserService {
         }
         dto.setEmail(user.getEmail());
         dto.setRole(user.getRole());
-        dto.setCreatedAt(user.getCreatedAt());
+        dto.setPassword(user.getPassword());
         dto.setCreatedBy(user.getCreatedBy());
         dto.setModifiedBy(user.getModifiedBy());
-        dto.setModifiedAt(user.getModifiedAt());
-        dto.setPassword(user.getPassword());
-
         return dto;
     }
 
-    public UserDto createUser(UserDto dto) {
+
+    public UserDto createUser(UserDto dto)throws DuplicateResourceException {
 
         UserEntity user = dtoToEntity(dto);
         user.setPassword(passwordEncoder.encode(user.getPassword()));
         user.setRole(user.getRole().toUpperCase());
+        if (userRepo.existsByUsername(user.getUsername())){
+            throw new DuplicateResourceException("UserName already exists");
+        }
+        if (userRepo.existsByEmail(user.getEmail())){
+            throw new DuplicateResourceException("Email already exists");
+        }
         UserEntity saved = userRepo.save(user);
+
         saved.setCreatedBy(saved.getId());
         saved.setModifiedBy(saved.getId());
-        saved = userRepo.save(saved);
 
+        saved = userRepo.save(saved);
 
         return entityToDto(saved);
     }
 
-public UserDto updateUser(Long id, UserDto dto) {
-    UserEntity update = userRepo.findById(id)
-            .orElseThrow(() ->
-                    new ResourceNotFoundException("User not Found with this id " + id));
 
-    update.setUsername(dto.getUsername());
-    update.setEmail(dto.getEmail());
+    public UserDto updateUser(Long id, UserDto dto) throws UnauthorizedActionException {
 
-    if (dto.getPassword() != null && !dto.getPassword().isEmpty()) {
-        update.setPassword(passwordEncoder.encode(dto.getPassword()));
-    }
+        UserEntity update = userRepo.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found with id " + id));
 
-    update.setRole(dto.getRole().toUpperCase());
-    Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         String tokenUsername = auth.getName();
-        log.debug("Logged in username: {}",tokenUsername);
 
-        Long loggedInUserId = userRepo.findIdByUsername(tokenUsername).orElseThrow(null);
+        UserEntity currentUser = userRepo.findByUsername(tokenUsername)
+                .orElseThrow(() -> new ResourceNotFoundException("Logged-in user not found"));
 
-        if(loggedInUserId!=null){
-            update.setModifiedBy(loggedInUserId);
-            if(!loggedInUserId.equals(update.getId())){
-                log.warn("Unauthorized update attempt by user ID: {}", loggedInUserId);
-                try {
-                    throw new UnauthorizedActionException(
-                            "Only Logged-In User can update their own profile details."
-                    );
-                } catch (UnauthorizedActionException e) {
-                    throw new RuntimeException(e);
-                }
-            }
+
+        if (!update.getCreatedBy().equals(currentUser.getId())) {
+            log.warn("Unauthorized access by user {}", currentUser.getId());
+            throw new UnauthorizedActionException("You can update only your own profile");
         }
-        log.debug("username: {}, loggedInUserId: {}", tokenUsername, loggedInUserId);
+
+        update.setUsername(dto.getUsername());
         update.setName(dto.getFirstName()+" "+dto.getLastName());
         update.setEmail(dto.getEmail());
+        update.setRole(dto.getRole().toUpperCase());
 
-    UserEntity currentUser = userRepo.findByUsername(tokenUsername)
-            .orElseThrow(() -> new RuntimeException("Logged-in user not found"));
-    update.setModifiedBy(currentUser.getId());
+        if (dto.getPassword() != null && !dto.getPassword().isBlank()) {
+            update.setPassword(passwordEncoder.encode(dto.getPassword()));
+        }
 
-    return entityToDto(userRepo.save(update));
-}
-    public List<UserDto> getAllUsers() {
+        update.setModifiedBy(currentUser.getId());
+
+        return entityToDto(userRepo.save(update));
+    }
+
+
+    public List<UserDto> getAllUsers() throws UnauthorizedActionException {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = auth.getName();
+
+        UserEntity currentUser = userRepo.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("Logged-in user not found"));
+
+        if (!"ADMIN".equalsIgnoreCase(currentUser.getRole())) {
+            throw new UnauthorizedActionException("Only admin can see this detail");
+        }
 
         return userRepo.findAll()
                 .stream()
@@ -121,21 +133,67 @@ public UserDto updateUser(Long id, UserDto dto) {
                 .collect(Collectors.toList());
     }
 
-    public UserDto getUserById(Long id) {
+    public UserDto getUserById(Long id) throws UnauthorizedActionException {
+        UserEntity current = userRepo.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found with id " + id));
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = auth.getName();
 
-        return entityToDto(userRepo.findById(id).orElseThrow(()->new RuntimeException("User not found")));
+        UserEntity currentUser = userRepo.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("Logged-in user not found"));
+
+        if (!current.getCreatedBy().equals(currentUser.getId()) &&
+                !"ADMIN".equalsIgnoreCase(currentUser.getRole())&&
+                !"OWNER".equalsIgnoreCase(currentUser.getRole())
+        ) {
+            throw new UnauthorizedActionException("Only created user can see their profile.");
+        }
+
+        return entityToDto(
+                userRepo.findById(id)
+                        .orElseThrow(() ->
+                                new ResourceNotFoundException("User not found with id " + id))
+        );
     }
+
+
 
     public ResponseEntity<?> deleteUser(Long id) {
+        UserEntity delete = userRepo.findById(id)
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found with id " + id));
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String tokenUsername = auth.getName();
+
+        UserEntity currentUser = userRepo.findByUsername(tokenUsername)
+                .orElseThrow(() -> new ResourceNotFoundException("Logged-in user not found"));
+
+
+        if (!delete.getCreatedBy().equals(currentUser.getId())) {
+            log.warn("Unauthorized access by user {}", currentUser.getId());
+            try {
+                throw new UnauthorizedActionException("Only created user can delete their profile");
+            } catch (UnauthorizedActionException e) {
+                throw new RuntimeException(e);
+            }
+        }
         userRepo.deleteById(id);
-
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("User deleted");
     }
-    public ResponseEntity<?> deleteAllUsers() {
+
+    public ResponseEntity<?> deleteAllUsers() throws UnauthorizedActionException {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        String username = auth.getName();
+
+        UserEntity currentUser = userRepo.findByUsername(username)
+                .orElseThrow(() -> new ResourceNotFoundException("Logged-in user not found"));
+
+        if (!"OWNER".equalsIgnoreCase(currentUser.getRole())) {
+            throw new UnauthorizedActionException("Only admin can see this detail");
+        }
         userRepo.deleteAll();
-
-        return ResponseEntity.ok().build();
+        return ResponseEntity.ok("All users deleted");
     }
-
-
 }
